Added sight and organized AI features
Added the ability for enemies to only move if they see the player. This is for when the enemies should not be able to see the player and therefore should not move.
Moved room activation into a new script in order to make the AI of skeleton and bat more OOP friendly.